pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = "Minimal FastAPI backend"
authors = ["mattias@archleap.se>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.95.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"